+++ huey/storage.py
@@ -817,6 +817,8 @@ class BaseSqlStorage(BaseStorage):
 
 class SqliteStorage(BaseSqlStorage):
     begin_sql = 'begin exclusive'
+    integrity_error = sqlite3.IntegrityError
+    sqlite_version_info = sqlite3.sqlite_version_info
     table_kv = ('create table if not exists kv ('
                 'queue text not null, key text not null, value blob not null, '
                 'primary key(queue, key))')
@@ -838,7 +840,7 @@ class SqliteStorage(BaseSqlStorage):
            table_counter]
 
     def __init__(self, name='huey', filename='huey.db', cache_mb=8,
-                 fsync=None, journal_mode='wal', timeout=5, strict_fifo=False,
+                 fsync=False, journal_mode='wal', timeout=5, strict_fifo=False,
                  create_tables=True, **kwargs):
         self.filename = filename
         self._cache_mb = cache_mb
@@ -954,7 +956,7 @@ class SqliteStorage(BaseSqlStorage):
 
     def pop_data(self, key):
         with self.db(commit=True) as curs:
-            if sqlite3.sqlite_version_info >= (3, 35, 0):
+            if self.sqlite_version_info >= (3, 35, 0):
                 curs.execute('delete from kv where queue = ? and key = ? '
                              'returning value', (self.name, key))
                 result = curs.fetchone()
@@ -981,21 +983,21 @@ class SqliteStorage(BaseSqlStorage):
                 curs.execute('insert or abort into kv '
                              '(queue, key, value) values (?, ?, ?)',
                              (self.name, key, self.to_blob(value)))
-        except sqlite3.IntegrityError:
+        except self.integrity_error:
             return False
         else:
             return True
 
     def incr(self, key, amount=1):
         with self.db(commit=True) as curs:
-            if sqlite3.sqlite_version_info >= (3, 35, 0):
+            if self.sqlite_version_info >= (3, 35, 0):
                 curs.execute('insert into counter (queue, key, value) '
                              'values (?, ?, ?) on conflict (queue, key) '
                              'do update set value = value + ? '
                              'returning value',
                              (self.name, key, amount, amount))
                 value, = curs.fetchone()
-            elif sqlite3.sqlite_version_info >= (3, 24, 0):
+            elif self.sqlite_version_info >= (3, 24, 0):
                 curs.execute('insert into counter (queue, key, value) '
                              'values (?, ?, ?) on conflict (queue, key) '
                              'do update set value = value + ?',
@@ -1031,23 +1033,27 @@ class SqliteStorage(BaseSqlStorage):
 
 class CySqliteStorage(SqliteStorage):
     def __init__(self, name='huey', filename='huey.db', pragmas=None,
-                 strict_fifo=False, create_tables=True, **kwargs):
+                 timeout=5, strict_fifo=False, create_tables=True, **kwargs):
         if cysqlite is None:
             raise ConfigurationError('"cysqlite" not found. Run "pip install '
                                      'cysqlite" to install.')
+        self.integrity_error = cysqlite.IntegrityError
+        self.sqlite_version_info = cysqlite.sqlite_version_info
+
         # Normalize hard-coded params to generic pragmas.
-        pragmas = pragmas or {}
+        pragmas = dict(pragmas or {})
         pragmas.setdefault('journal_mode', 'wal')
         if 'journal_mode' in kwargs:
             pragmas['journal_mode'] = kwargs.pop('journal_mode') or 'wal'
         if 'cache_mb' in kwargs:
             pragmas['cache_size'] = kwargs.pop('cache_mb') * -1000
         if 'fsync' in kwargs:
-            pragmas['synchronous'] = 2 if kwargs.pop('synchronous') else 0
+            pragmas['synchronous'] = 2 if kwargs.pop('fsync') else 0
 
         super(CySqliteStorage, self).__init__(
             name,
             filename,
+            timeout=timeout,
             strict_fifo=strict_fifo,
             create_tables=create_tables,
             pragmas=pragmas,
@@ -1056,17 +1062,6 @@ class CySqliteStorage(SqliteStorage):
     def _create_connection(self):
         return cysqlite.connect(self.filename, timeout=self._timeout,
                                 **self._conn_kwargs)
-
-    def put_if_empty(self, key, value):
-        try:
-            with self.db(commit=True) as curs:
-                curs.execute('insert or abort into kv '
-                             '(queue, key, value) values (?, ?, ?)',
-                             (self.name, key, self.to_blob(value)))
-        except cysqlite.IntegrityError:
-            return False
-        else:
-            return True
 
 
 class PostgresStorage(BaseSqlStorage):